% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficarSegmento.r
\name{graficarSegmento}
\alias{graficarSegmento}
\title{graficarSegmento
Grafica un segmento dado 2 puntos}
\usage{
graficarSegmento(a, b)
}
\arguments{
\item{a}{vector}

\item{b}{vector}
}
\value{
vector
}
\description{
graficarSegmento
Grafica un segmento dado 2 puntos
}
\examples{
graficarSegmento(c(1,1),c(2,2))
}
