% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puntoMedio.r
\name{puntoMedio}
\alias{puntoMedio}
\title{punto medio
Calcula el punto que se encuentra entre 2 puntos}
\usage{
puntoMedio(a, b)
}
\arguments{
\item{a}{vector}

\item{b}{vector}
}
\value{
vector
}
\description{
punto medio
Calcula el punto que se encuentra entre 2 puntos
}
\examples{
puntoMedio(c(1,1),c(2,2))
}
